+++ pages/public.php
@@ -1,0 +1,5 @@
+<?php
+
+global $template_dir;
+
+include $template_dir . '/pages/cache.php';